Updating location and forecast details
Calls Google Geocode to get location city, state, latitude, and longitude. Lat/long are then used in calling the openWeatherAPI. This ensures that the correct location is called and returned. In using city, state, and country code would get results that were incorrect from the openWeatherAPI. In using lat/long the results are much more correct and desirable.